package rename correction
