+++ app/server/appsmith-server/src/main/java/com/appsmith/server/models/export/PageMetadata.java
@@ -1,4 +1,4 @@
-package com.appsmith.server.modals;
+package com.appsmith.server.models.export;
 
 import java.util.List;
